requirements.txt
python-dotenv
eth-tester
flask>=3.0.0
flask-cors
gunicorn
py-evm
web3==7.9.0